refactor：移动凭证文件管道至pipes.py
